pyproject.toml
[build-system]
requires = ["setuptools>=61,<80"]
build-backend = "setuptools.build_meta"

[project]
name = "llm4netlab"
version = "0.1.0"
description = ""
authors = [{ name = "zhihao1998", email = "zhihao.wang@polito.it" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic",
    "openai",
    "instructor",
    "kathara>=3.7.9",
    "langfuse",
    "python-dotenv",
    "langchain-deepseek",
    "langgraph",
    "mcp[cli]",
    "mcp-use",
    "langchain-mcp-adapters",
    "google-api-python-client",
    "setuptools>=61,<80",
    "fs>=2.4.16",
    "langchain-ollama",
    "langchain==1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["llm4netlab*", "agent*"]

